crewai-crash-course-modified/requirements.txt
# Core CrewAI dependencies
crewai[tools]>=0.157.0
crewai-tools>=0.60.0

# LLM and API dependencies
litellm>=1.74.9
openai>=1.12.0
google-generativeai>=0.8.0

# Environment and configuration
python-dotenv>=1.0.0

# Data processing and utilities
pydantic>=2.6.1
requests>=2.31.0

# Memory and storage
chromadb>=0.5.23
lancedb>=0.5.4

# Web scraping and tools
beautifulsoup4>=4.12.2
pytube>=15.0.0

# File handling
openpyxl>=3.1.5
pypdf>=5.0.0
pdfplumber>=0.11.4

# Development and utilities
uv>=0.4.25
